+++ packages/engine-server/src/migrations/migrations.ts
@@ -1,3 +1,4 @@
+import { ScratchConfig } from "@dendronhq/common-all";
 import {
   SegmentClient,
   TelemetryStatus,
@@ -8,14 +9,17 @@ import { DConfig } from "../config";
 import { removeCache } from "../utils";
 import { Migrations } from "./types";
 
+/**
+ * Migrations are sorted by version numbers, from greatest to least
+ */
 export const ALL_MIGRATIONS: Migrations[] = [
   {
-    version: "0.52.0",
+    version: "0.51.4",
     changes: [
       {
         name: "migrate scratch config",
         func: async ({ dendronConfig, wsConfig }) => {
-          dendronConfig.scratch = DConfig.genDefaultConfig().scratch;
+          dendronConfig.scratch = DConfig.genDefaultConfig().scratch as ScratchConfig;
           if (_.get(wsConfig.settings, "dendron.defaultScratchName")) {
             dendronConfig.scratch.name = _.get(
               wsConfig.settings,
@@ -33,6 +37,21 @@ export const ALL_MIGRATIONS: Migrations[] = [
               wsConfig.settings,
               "dendron.defaultScratchAddBehavior"
             );
+          }
+          return { data: { dendronConfig, wsConfig } };
+        },
+      },
+    ],
+  },
+  {
+    version: "0.51.4",
+    changes: [
+      {
+        name: "don't switch to legacy preview if not currently on it",
+        func: async ({ dendronConfig, wsConfig }) => {
+          const previewV2Enabled = dendronConfig.dev?.enablePreviewV2
+          if (!previewV2Enabled) {
+            _.set(dendronConfig, "dev.previewV2Enabled", false)
           }
           return { data: { dendronConfig, wsConfig } };
         },