+++ packages/plugin-core/src/commands/NoteLookupCommand.ts
@@ -11,7 +11,10 @@ import { Uri } from "vscode";
 import {
   DirectChildFilterBtn,
   JournalBtn,
+  ScratchBtn,
   MultiSelectBtn,
+  Selection2LinkBtn,
+  SelectionExtractBtn,
 } from "../components/lookup/buttons";
 import { LookupControllerV3 } from "../components/lookup/LookupControllerV3";
 import {
@@ -34,6 +37,8 @@ import {
   LookupFilterType,
   LookupNoteType,
   LookupNoteTypeEnum,
+  LookupSelectionType,
+  LookupSelectionTypeEnum,
 } from "./LookupCommand";
 
 type CommandRunOpts = {
@@ -42,6 +47,7 @@ type CommandRunOpts = {
   fuzzThreshold?: number;
   multiSelect?: boolean;
   noteType?: LookupNoteType;
+  selectionType?: LookupSelectionType;
   /**
    * NOTE: currently, only one filter is supported
    */
@@ -66,7 +72,7 @@ type CommandOpts = {
   selectedItems: readonly NoteQuickInput[];
 } & CommandGatherOutput;
 
-type CommandOutput = {
+export type CommandOutput = {
   quickpick: DendronQuickPickerV2;
   controller: LookupControllerV3;
   provider: ILookupProviderV3;
@@ -130,11 +136,15 @@ export class NoteLookupCommand extends BaseCommand<
         "lookupConfirmVaultOnCreate"
       ),
       extraButtons: [
+        //todo: mirror v2 button sequence
         JournalBtn.create(copts.noteType === LookupNoteTypeEnum.journal),
+        ScratchBtn.create(copts.noteType === LookupNoteTypeEnum.scratch),
         MultiSelectBtn.create(copts.multiSelect),
         DirectChildFilterBtn.create(
           copts.filterMiddleware?.includes("directChildOnly")
         ),
+        Selection2LinkBtn.create(copts.selectionType === LookupSelectionTypeEnum.selection2link),
+        SelectionExtractBtn.create(copts.selectionType === LookupSelectionTypeEnum.selectionExtract),
       ],
     });
     this._provider = new NoteLookupProvider("lookup", {
@@ -267,6 +277,9 @@ export class NoteLookupCommand extends BaseCommand<
         ? picker.vault
         : PickerUtilsV2.getOrPromptVaultForOpenEditor();
       nodeNew = NoteUtils.create({ fname, vault });
+      if (picker.selectionProcessFunc !== undefined) {
+        nodeNew = await picker.selectionProcessFunc(nodeNew) as NoteProps;
+      }
     }
     const resp = await engine.writeNote(nodeNew, {
       newNode: true,