+++ packages/plugin-core/src/test/suite-integ/Extension.test.ts
@@ -1,4 +1,4 @@
-import { CONSTANTS, isNotUndefined, Time } from "@dendronhq/common-all";
+import { CONSTANTS, DendronConfig, isNotUndefined, Time } from "@dendronhq/common-all";
 import { readYAML, tmpDir } from "@dendronhq/common-server";
 import { getPortFilePath, getWSMetaFilePath, MetadataService, openWSMetaFile } from "@dendronhq/engine-server";
 import { TestEngineUtils } from "@dendronhq/engine-test-utils";
@@ -121,6 +121,9 @@ suite("Extension", function () {
               noXVaultWikiLink: true,
               lookupConfirmVaultOnCreate: false,
               autoFoldFrontmatter: true,
+              dev: {
+                enablePreviewV2: true,
+              },
               mermaid: true,
               useKatex: true,
               site: {
@@ -135,7 +138,7 @@ suite("Extension", function () {
                   payload: ["vault"],
                 },
               },
-            });
+            } as DendronConfig);
             const dendronState = MetadataService.instance().getMeta();
             expect(isNotUndefined(dendronState.firstInstall)).toBeTruthy();
             expect(isNotUndefined(dendronState.firstWsInitialize)).toBeTruthy();