+++ packages/common-test-utils/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.52.0](https://github.com/dendronhq/dendron/compare/v0.51.3...v0.52.0) (2021-07-26)
+
+**Note:** Version bump only for package @dendronhq/common-test-utils
+
+
+
+
+
+## [0.51.4](https://github.com/dendronhq/dendron/compare/v0.51.3...v0.51.4) (2021-07-25)
+
+**Note:** Version bump only for package @dendronhq/common-test-utils
+
+
+
+
+
 ## [0.51.2](https://github.com/dendronhq/dendron/compare/v0.51.0...v0.51.2) (2021-07-22)
 
 **Note:** Version bump only for package @dendronhq/common-test-utils